update examples notebook, readme files, rename custom provider files
