--- .env.example
@@ -1,109 +1,0 @@
-# Example Environment Configuration
-# Copy this file and update with your actual values
-
-# Application Settings
-APP_NAME=MMCT Agent
-APP_VERSION=1.0.0
-ENVIRONMENT=development
-DEBUG=false
-
-# LLM Provider Configuration
-LLM_PROVIDER=azure
-LLM_USE_MANAGED_IDENTITY=true
-LLM_TIMEOUT=200
-LLM_MAX_RETRIES=2
-LLM_TEMPERATURE=0.0
-LLM_ENDPOINT=https://your-openai-endpoint.openai.azure.com/
-LLM_DEPLOYMENT_NAME=gpt-4o
-LLM_MODEL_NAME=gpt-4o
-LLM_API_VERSION=2024-08-01-preview
-LLM_API_KEY=your-api-key-if-not-using-managed-identity-or-using-openai-directly
-LLM_VISION_DEPLOYMENT_NAME=gpt-4o
-LLM_VISION_API_VERSION=2024-08-01-preview
-
-# Search Provider Configuration
-# Search Provider Options: azure_ai_search, local_faiss
-SEARCH_PROVIDER=azure_ai_search
-SEARCH_USE_MANAGED_IDENTITY=true
-SEARCH_TIMEOUT=30
-SEARCH_ENDPOINT=https://your-search-service.search.windows.net
-SEARCH_API_KEY=your-search-api-key-if-not-using-managed-identity
-
-# Embedding Provider Configuration
-EMBEDDING_PROVIDER=azure
-EMBEDDING_USE_MANAGED_IDENTITY=true
-EMBEDDING_TIMEOUT=200
-EMBEDDING_SERVICE_ENDPOINT=https://your-openai-endpoint.openai.azure.com/
-EMBEDDING_SERVICE_DEPLOYMENT_NAME=text-embedding-ada-002
-EMBEDDING_SERVICE_API_VERSION=2024-08-01-preview
-EMBEDDING_SERVICE_MODEL_NAME=text-embedding-ada-002
-EMBEDDING_SERVICE_API_KEY=your-embedding-api-key-if-not-using-managed-identity-or-using-openai-directly
-
-# Transcription Provider Configuration
-TRANSCRIPTION_PROVIDER=azure
-SPEECH_USE_MANAGED_IDENTITY=true
-
-## Azure STT
-SPEECH_TIMEOUT=200
-SPEECH_SERVICE_REGION=centralindia
-SPEECH_SERVICE_RESOURCE_ID=/subscriptions/your-subscription-id/resourceGroups/your-rg/providers/Microsoft.CognitiveServices/accounts/your-speech-service
-SPEECH_SERVICE_ENDPOINT=https://centralindia.api.cognitive.microsoft.com/
-
-## Azure OpenAI's Whisper
-SPEECH_SERVICE_DEPLOYMENT_NAME=whisper
-SPEECH_SERVICE_MODEL_NAME=whisper
-SPEECH_SERVICE_API_VERSION=2024-08-01-preview
-WHISPER_ENDPOINT=https://your-openai-endpoint.openai.azure.com/
-
-### Key for Speech (Compatible for Azure STT, Azure OpenAI, and OpenAI)
-SPEECH_SERVICE_KEY=your-speech-api-key-if-not-using-managed-identity-or-using-openai-directly
-
-# Vision Provider Configuration
-VISION_PROVIDER=azure
-
-# Storage Configuration
-# Storage Provider Options: azure, local
-STORAGE_PROVIDER=azure
-STORAGE_USE_MANAGED_IDENTITY=true
-STORAGE_CONTAINER_NAME=default
-STORAGE_ACCOUNT_NAME=your-storage-account-name
-STORAGE_CONNECTION_STRING=your-storage-connection-string-if-not-using-managed-identity
-STORAGE_ACCOUNT_URL=https://your-storage-account-name.blob.core.windows.net/
-BLOB_CONNECTION_STRING=your-blob-connection-string-if-not-using-managed-identity
-BLOB_SAS_TOKEN=your-blob-sas-token-if-needed
-
-# Storage Container Configuration
-VIDEO_CONTAINER_NAME=mmct-videocontainer
-FRAMES_CONTAINER_NAME=mmct-framescontainer
-TIMESTAMPS_CONTAINER_NAME=mmct-timestampscontainer
-TRANSCRIPT_CONTAINER_NAME=mmct-transcriptcontainer
-AUDIO_CONTAINER_NAME=mmct-audiocontainer
-VIDEO_DESCRIPTION_CONTAINER_NAME=mmct-summary-n-transcript
-BLOB_DOWNLOAD_DIR=media
-BLOB_MANAGED_IDENTITY=true
-
-
-# Logging Configuration
-LOG_LEVEL=INFO
-LOG_FILE=logs/mmct-dev.log
-LOG_ENABLE_JSON=false
-LOG_ENABLE_FILE=true
-LOG_MAX_FILE_SIZE=10 MB
-LOG_RETENTION_DAYS=7
-
-
-# We set TOKENIZERS_PARALLELISM=false to avoid deadlock warnings from Hugging Face tokenizers.
-# This disables Rust‑based parallel tokenization, which can hang when using multiprocessing.
-#
-# If you’re NOT using multiprocessing (e.g. num_workers = 1), you can safely set it to true for faster tokenization.
-TOKENIZERS_PARALLELISM=false
-# enable INFO‑level logs from 🤗 Transformers.
-TRANSFORMERS_VERBOSITY=info
-
-# TRANSFORMERS_VERBOSITY controls the log level for Hugging Face Transformers.
-# Allowed values:
-#   - debug     : Very detailed logs, useful for troubleshooting internals.
-#   - info      : General operational logs (default-level informative output).
-#   - warning   : Only show warnings and errors.
-#   - error     : Show only errors.
-#   - critical  : Only the most critical issues.
