+++ mmct/providers/custom_providers/__init__.py
@@ -1,7 +1,7 @@
-from .search_provider import CustomSearchProvider
+from .graph_rag_search_provider import CustomSearchProvider
 from .local_faiss_search_provider import LocalFaissSearchProvider
-from .image_embedding_provider import CustomImageEmbeddingProvider
-from .storage_provider import LocalStorageProvider
+from .clip_image_embedding_provider import CustomImageEmbeddingProvider
+from .local_storage_provider import LocalStorageProvider
 
 __all__ = [
     'CustomSearchProvider',
