+++ mmct/providers/custom_providers/clip_image_embedding_provider.py
